backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
Django==6.0.1
django-htmx==1.27.0
djangorestframework==3.16.1
faiss-cpu==1.13.2
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.2
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.1
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.4
psycopg2-binary==2.9.11
py-ubjson==0.16.1
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.5.0
pyreadline3==3.5.4
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.12.0
whitenoise==6.11.0
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.22.0
zope.interface==8.2
zstandard==0.25.0
razorpay==1.4.2

